Add files via upload
Created in 2022 using AutoHotkey (AHK), this program automates updates to a specific database (DB) for an employee who doesn't have direct access to it due to restrictions associated with their position.
The updates are performed based on information provided in an Excel file. The program automates the process by copying data from the Excel and pasting it into the database's graphical user interface (GUI), streamlining the workflow between Excel and the database program.

Note: This program is no longer maintained, and no further changes or updates will be made.